Move the new code into a separate file
